Added terrain generation

+++ Driving-Game/Code/PlayerData.cs
@@ -4,7 +4,7 @@ public class PlayerData
     public int Rotation { get; set; }
     public int Slope { get; set; }
     public int DistToGround { get; set; }
-    public double AngularVelocity { get; set; }
+    public int AngularVelocity { get; set; }
     public bool IsTouchingGround { get; set; }
     public bool HasDied { get; set; }
     public string Print()
